refactor(#1074): extract shared JsonApiResponseTrait
HttpKernel and ControllerDispatcher had identical private
jsonApiResponse() methods. Extract to a shared trait in
Waaseyaa\Foundation\Http\JsonApiResponseTrait.
