+++ packages/foundation/src/Http/ControllerDispatcher.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Waaseyaa\Foundation\Http;
 
 use Symfony\Component\HttpFoundation\File\UploadedFile;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request as HttpRequest;
 use Symfony\Component\HttpFoundation\Response as HttpResponse;
 use Symfony\Component\HttpFoundation\StreamedResponse;
@@ -44,6 +43,8 @@ use Waaseyaa\SSR\SsrPageHandler;
  */
 final class ControllerDispatcher
 {
+    use JsonApiResponseTrait;
+
     private readonly LoggerInterface $logger;
 
     public function __construct(
@@ -986,20 +987,6 @@ final class ControllerDispatcher
         return '/files/' . ltrim($path, '/');
     }
 
-    /**
-     * Build a JSON:API response with correct content type and encoding.
-     *
-     * @param array<string, mixed> $data
-     * @param array<string, string> $headers
-     */
-    private function jsonApiResponse(int $status, array $data, array $headers = []): JsonResponse
-    {
-        $response = new JsonResponse($data, $status, $headers);
-        $response->setEncodingOptions(JSON_UNESCAPED_SLASHES | JSON_PRETTY_PRINT | JSON_THROW_ON_ERROR);
-        $response->headers->set('Content-Type', 'application/vnd.api+json');
-
-        return $response;
-    }
 
     /**
      * Build an HTML response.
